style: bignumber charts have smaller tiles by default
