+++ packages/common/src/types/savedCharts.ts
@@ -1,0 +1,9 @@
+// eslint-disable-next-line import/prefer-default-export
+export enum DBChartTypes {
+    COLUMN = 'column',
+    BAR = 'bar',
+    LINE = 'line',
+    SCATTER = 'scatter',
+    TABLE = 'table',
+    BIG_NUMBER = 'big_number',
+}
